requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
aioboto3==13.3.0
    # via -r requirements.in
aiobotocore[boto3]==2.16.0
    # via
    #   aioboto3
    #   aiobotocore
aiofiles==24.1.0
    # via aioboto3
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via aiobotocore
aioitertools==0.12.0
    # via aiobotocore
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
attrs==25.3.0
    # via aiohttp
boto3==1.35.81
    # via aiobotocore
botocore==1.35.81
    # via
    #   aiobotocore
    #   boto3
    #   s3transfer
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
click==8.3.0
    # via uvicorn
fastapi==0.115.14
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.25.2
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   yarl
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
json-log-formatter==1.1.1
    # via -r requirements.in
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.9
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.1.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
redis==5.0.8
    # via -r requirements.in
s3transfer==0.10.4
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.46.2
    # via fastapi
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via botocore
uvicorn==0.23.2
    # via -r requirements.in
wrapt==1.17.3
    # via aiobotocore
yarl==1.20.1
    # via aiohttp
